src/main/webapp/app/entities/diplomes/route/diplomes-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IDiplomes, Diplomes } from '../diplomes.model';
import { DiplomesService } from '../service/diplomes.service';

import { DiplomesRoutingResolveService } from './diplomes-routing-resolve.service';

describe('Diplomes routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: DiplomesRoutingResolveService;
  let service: DiplomesService;
  let resultDiplomes: IDiplomes | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(DiplomesRoutingResolveService);
    service = TestBed.inject(DiplomesService);
    resultDiplomes = undefined;
  });

  describe('resolve', () => {
    it('should return IDiplomes returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDiplomes = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDiplomes).toEqual({ id: 123 });
    });

    it('should return new IDiplomes if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDiplomes = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultDiplomes).toEqual(new Diplomes());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Diplomes })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDiplomes = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDiplomes).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
